# doc/api.yaml
openapi: 3.0.0
paths:
...
  /artist:
    get:
      description: Retrieve a list of all artists
      operationId: ArtistsController_getAllArtists
      parameters: []
      responses:
        "200":
          description: List of artists retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artist"
      security: &a3
        - bearer: []
      summary: Get all artists
      tags: &a4
        - Artists
    post:
      description: Create a new artist with name and grammy award status
      operationId: ArtistsController_createArtist
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtistDto"
      responses:
        "201":
          description: Artist created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: Bad request - invalid input data
      security: *a3
      summary: Create a new artist
      tags: *a4
  /artist/{id}:
    get:
      description: Retrieve a specific artist by their ID
      operationId: ArtistsController_getArtistById
      parameters:
        - name: id
          required: true
          in: path
          description: Artist UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "200":
          description: Artist retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: Invalid UUID format
        "404":
          description: Artist not found
      security: *a3
      summary: Get artist by ID
      tags: *a4
    put:
      description: Update artist information
      operationId: ArtistsController_updateArtist
      parameters:
        - name: id
          required: true
          in: path
          description: Artist UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArtistDto"
      responses:
        "200":
          description: Artist updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: Bad request - invalid input data
        "404":
          description: Artist not found
      security: *a3
      summary: Update artist
      tags: *a4
    delete:
      description: Delete an artist by their ID. This will also remove artist
        references from tracks and albums.
      operationId: ArtistsController_deleteArtist
      parameters:
        - name: id
          required: true
          in: path
          description: Artist UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "204":
          description: Artist deleted successfully
        "400":
          description: Invalid UUID format
        "404":
          description: Artist not found
      security: *a3
      summary: Delete artist
      tags: *a4
...
  /track/{id}:
    get:
      description: Retrieve a specific track by its ID
      operationId: TracksController_getTrackById
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "200":
          description: Track retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Invalid UUID format
        "404":
          description: Track not found
      security: *a5
      summary: Get track by ID
      tags: *a6
    put:
      description: Update track information
      operationId: TracksController_updateTrack
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrackDto"
      responses:
        "200":
          description: Track updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Bad request - invalid input data
        "404":
          description: Track not found
      security: *a5
      summary: Update track
      tags: *a6
    delete:
      description: Delete a track by its ID. This will also remove the track from
        favorites if present.
      operationId: TracksController_deleteTrack
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "204":
          description: Track deleted successfully
        "400":
          description: Invalid UUID format
        "404":
          description: Track not found
      security: *a5
      summary: Delete track
      tags: *a6
...
  /album/{id}:
    get:
      operationId: AlbumsController_getAlbumById
      parameters:
        - name: id
          required: true
          in: path
          description: Album ID
          schema:
            type: string
      responses:
        "200":
          description: Return album by ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: Invalid UUID
        "404":
          description: Album not found
      summary: Get album by ID
      tags: *a7
    put:
      operationId: AlbumsController_updateAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: Album ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlbumDto"
      responses:
        "200":
          description: Album updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: Bad request
        "404":
          description: Album not found
      summary: Update album
      tags: *a7
    delete:
      operationId: AlbumsController_deleteAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: Album ID
          schema:
            type: string
      responses:
        "204":
          description: Album deleted
        "400":
          description: Invalid UUID
        "404":
          description: Album not found
      summary: Delete album
      tags: *a7
  /favs:
    get:
      description: Retrieve all favorite artists, albums, and tracks
      operationId: FavoritesController_getAllFavorites
      parameters: []
      responses:
        "200":
          description: Favorites retrieved successfully
          content:
            application/json:
              schema:
                type: object
      security: &a8
        - bearer: []
      summary: Get all favorites
      tags: &a9
        - Favorites
  /favs/track/{id}:
    post:
      description: Add a track to favorites by its ID
      operationId: FavoritesController_addTrackToFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "201":
          description: Track added to favorites successfully
        "400":
          description: Invalid UUID format
        "422":
          description: Track not found - cannot add to favorites
      security: *a8
      summary: Add track to favorites
      tags: *a9
    delete:
      description: Remove a track from favorites by its ID
      operationId: FavoritesController_removeTrackFromFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "204":
          description: Track removed from favorites successfully
        "400":
          description: Invalid UUID format
        "404":
          description: Track not found in favorites
      security: *a8
      summary: Remove track from favorites
      tags: *a9
  /favs/album/{id}:
    post:
      description: Add an album to favorites by its ID
      operationId: FavoritesController_addAlbumToFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: Album UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "201":
          description: Album added to favorites successfully
        "400":
          description: Invalid UUID format
        "422":
          description: Album not found - cannot add to favorites
      security: *a8
      summary: Add album to favorites
      tags: *a9
    delete:
      description: Remove an album from favorites by its ID
      operationId: FavoritesController_removeAlbumFromFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: Album UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "204":
          description: Album removed from favorites successfully
        "400":
          description: Invalid UUID format
        "404":
          description: Album not found in favorites
      security: *a8
      summary: Remove album from favorites
      tags: *a9
  /favs/artist/{id}:
    post:
      description: Add an artist to favorites by its ID
      operationId: FavoritesController_addArtistToFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: Artist UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "201":
          description: Artist added to favorites successfully
        "400":
          description: Invalid UUID format
        "422":
          description: Artist not found - cannot add to favorites
      security: *a8
      summary: Add artist to favorites
      tags: *a9
    delete:
      description: Remove an artist from favorites by its ID
      operationId: FavoritesController_removeArtistFromFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: Artist UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "204":
          description: Artist removed from favorites successfully
        "400":
          description: Invalid UUID format
        "404":
          description: Artist not found in favorites
      security: *a8
      summary: Remove artist from favorites
      tags: *a9
info:
  title: Home Library Service
  description: Home music library service
  version: 1.0.0
  contact: {}
...
servers: []
components:
  schemas:
...
    Artist:
      type: object
      properties:
        id:
          type: string
          description: Artist UUID
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          description: Artist name
          example: Michael Jackson
        grammy:
          type: boolean
          description: Whether the artist has won a Grammy award
          example: true
      required:
        - id
        - name
        - grammy
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
          description: Artist name
          example: Michael Jackson
        grammy:
          type: boolean
          description: Whether the artist has won a Grammy award
          example: true
      required:
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
          description: Artist name
          example: Michael Jackson
        grammy:
          type: boolean
          description: Whether the artist has won a Grammy award
          example: true
    Track:
      type: object
      properties:
        id:
          type: string
          description: Track UUID
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          description: Track name
          example: Bohemian Rhapsody
        artistId:
          type: object
          description: Artist ID
          example: 123e4567-e89b-12d3-a456-426614174000
          nullable: true
        albumId:
          type: object
          description: Album ID
          example: 123e4567-e89b-12d3-a456-426614174000
          nullable: true
        duration:
          type: number
          description: Track duration in seconds
          example: 354
      required:
        - id
        - name
        - duration
...
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
          description: Track name
          example: Bohemian Rhapsody
        artistId:
          type: object
          description: Artist ID (optional)
          example: 123e4567-e89b-12d3-a456-426614174000
          nullable: true
        albumId:
          type: object
          description: Album ID (optional)
          example: 123e4567-e89b-12d3-a456-426614174000
          nullable: true
        duration:
          type: number
          description: Track duration in seconds
          example: 354
          minimum: 1
    Album:
      type: object
      properties:
        id:
          type: string
          description: Album ID
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          description: Album name
          example: Thriller
        year:
          type: number
          description: Release year
          example: 1982
        artistId:
          type: object
          description: Artist ID
          example: 123e4567-e89b-12d3-a456-426614174000
          nullable: true
      required:
        - id
        - name
        - year
        - artistId
...
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
          description: Album name
          example: Thriller
        year:
          type: number
          description: Release year
          example: 1982
        artistId:
          type: object
          description: Artist ID
          example: 123e4567-e89b-12d3-a456-426614174000
          nullable: true